+++ src/snowflake/cli/api/project/schemas/entities/common.py
@@ -18,6 +18,7 @@ from abc import ABC
 from typing import Any, Dict, Generic, List, Optional, TypeVar, Union
 
 from pydantic import Field, PrivateAttr, field_validator
+from pydantic_core.core_schema import ValidationInfo
 from snowflake.cli.api.identifiers import FQN
 from snowflake.cli.api.project.schemas.updatable_model import (
     IdentifierField,
@@ -44,6 +45,42 @@ class SqlScriptHookType(UpdatableModel):
 PostDeployHook = SqlScriptHookType
 
 
+class Dependency(UpdatableModel):
+    entity_id: str = Field(title="Id of the entity", alias="id")
+    arguments: Optional[Dict[str, Union[int, bool, str]]] = Field(
+        title="Arguments that will be passed to entity build and deploy actions, provided as a key value pairs",
+        default_factory=dict,
+    )
+
+    @field_validator("arguments", mode="before")
+    @classmethod
+    def arguments_validator(cls, arguments: Dict, info: ValidationInfo) -> Dict:
+        duplicated_run = (
+            info.context.get("is_duplicated_run", False) if info.context else False
+        )
+        if not duplicated_run:
+            for k, v in arguments.items():
+                arguments[k] = cls._cast_value(v)
+
+        return arguments
+
+    @staticmethod
+    def _cast_value(value: str) -> Union[int, bool, str]:
+        if value.lower() in ["true", "false"]:
+            return value.lower() == "true"
+
+        try:
+            return int(value)
+        except ValueError:
+            return value
+
+    def __eq__(self, other):
+        return self.entity_id == other.entity_id
+
+    def __hash__(self):
+        return hash(self.entity_id)
+
+
 class MetaField(UpdatableModel):
     warehouse: Optional[str] = IdentifierField(
         title="Warehouse used to run the scripts", default=None
@@ -59,6 +96,10 @@ class MetaField(UpdatableModel):
     use_mixins: Optional[List[str]] = Field(
         title="Name of the mixin used to fill the entity fields",
         default=None,
+    )
+
+    depends_on: Optional[List[Dependency]] = Field(
+        title="Entities that need to be deployed before this one", default_factory=list
     )
 
     @field_validator("use_mixins", mode="before")